[Circuit-Breaker] Add memory threshold in opensearch-js client (#207)
This is to add memory checker to avoid dashboard out of memory crash.
There will be one field introduced in `ClientOptions`:  `memoryCircuitBreaker`. 
The default value of `memoryCircuitBreaker.enabled` is false until customer add manual config to modify it.
The default value of `memoryCircuitBreaker.maxPercentage` is `1.0` until customer modify it. The proper values should be within [0, 1] inclusively. 

Issues Resolved
https://github.com/opensearch-project/opensearch-js/issues/202

Signed-off-by: Zuocheng Ding <zding817@gmail.com>